test: move grouped_topk accuracy test to test_special_ops.py and perf test to test_special_perf.py
